requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
